st_ino should actually be uint64_t

+++ kernel/include/kernel/fs/stat.h
@@ -9,7 +9,7 @@ typedef struct {
 
 typedef struct {
 	short		   st_dev;
-	unsigned short st_ino;
+	uint64_t	   st_ino;
 	uint32_t	   st_mode;
 	unsigned short st_nlink;
 	unsigned short st_uid;
